fixed a typing bug in the algorithm causing infinite loops with certain inputs, changed the py test runner to be interactive

+++ cpp/src/algorithm.cpp
@@ -21,8 +21,8 @@ private:
 		while (true) {
 			// BELLMAN FORD
 			// distance source to each node
-			vector<int> dist(numNodes, __INT_MAX__);
-			dist[sourceIdx] = 0;
+			vector<double> dist(numNodes, __INT_MAX__);
+			dist[sourceIdx] = 0.0f;
 
 			// path tracking
 			vector<int> parentNode(numNodes, -1);
